+++ bin/remove.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_native_splash/flutter_native_splash.dart'
+    as flutter_native_splash;
+
+void main(List<String> arguments) {
+  flutter_native_splash.removeSplash();
+}